+++ dgindexer/src/test/scala/com/stratio/governance/agent/searcher/actors/utils/AdditionalBusinessUnitTest.scala
@@ -10,9 +10,9 @@ class AdditionalBusinessUnitTest extends FlatSpec {
 
   "method getBTTotalIndexationsubquery" should "be processed properly" in {
 
-    val result: String = additionalBusiness.getAdditionalBusinessTotalIndexationSubquery("dg_metadata", "business_assets", "business_assets_type", "quality")
+    val result: String = additionalBusiness.getAdditionalBusinessTotalIndexationSubquery("dg_metadata", "business_assets", "business_assets_type", "bpm_status", "quality")
     assertResult("select ba.id as id,ba.name as name,'' as alias,ba.description as description,'' as metadata_path,'GLOSSARY' as type,'BUSINESS_TERM' as subtype,ba.tenant,null as properties,true as active,ba.modified_at as discovered_at,ba.modified_at as modified_at " +
-      "from dg_metadata.business_assets as ba, dg_metadata.business_assets_type as bat where ba.business_assets_type_id = bat.id and bat.name='TERM' " +
+      "from dg_metadata.business_assets as ba, dg_metadata.business_assets_type as bat, dg_metadata.bpm_status as bas where ba.business_assets_type_id = bat.id and bat.name='TERM' and ba.business_assets_status_id = bas.id and bas.active = true " +
       "UNION " +
       "select id,name,'' as alias,description,'' as metadata_path,'QUALITY' as type,'RULES' as subtype, tenant,null as properties, active, modified_at as discovered_at, modified_at from dg_metadata.quality")(result)
 
@@ -20,18 +20,18 @@ class AdditionalBusinessUnitTest extends FlatSpec {
 
   "method getBTPartialIndexationSubquery1" should "be processed properly" in {
 
-    val result: String = additionalBusiness.getAdditionalBusinessPartialIndexationSubquery1("dg_metadata", "business_assets", "business_assets_type", "quality", 7)
+    val result: String = additionalBusiness.getAdditionalBusinessPartialIndexationSubquery1("dg_metadata", "business_assets", "business_assets_type", "bpm_status", "quality", 7)
     assertResult("SELECT '',ba.id,ba.modified_at,7 " +
-      "FROM dg_metadata.business_assets as ba, dg_metadata.business_assets_type as bat " +
-      "WHERE ba.business_assets_type_id = bat.id and bat.name='TERM' and ba.modified_at > ? UNION " +
+      "FROM dg_metadata.business_assets as ba, dg_metadata.business_assets_type as bat, dg_metadata.bpm_status as bas " +
+      "WHERE ba.business_assets_type_id = bat.id and bat.name='TERM' and ba.business_assets_status_id = bas.id and bas.active = true and ba.modified_at > ? UNION " +
       "SELECT metadata_path, id, modified_at, 8 FROM dg_metadata.quality WHERE modified_at > ?")(result)
 
   }
 
   "method getBTPartialIndexationSubquery2" should "be processed properly" in {
 
-    val result: String = additionalBusiness.getBusinessTermsPartialIndexationSubquery2("dg_metadata", "business_assets", "business_assets_type")
-    assertResult("select ba.id as id,ba.name as name,'' as alias,ba.description as description,'' as metadata_path,'GLOSSARY' as type,'BUSINESS_TERM' as subtype,ba.tenant,null as properties,true as active,ba.modified_at as discovered_at,ba.modified_at as modified_at from dg_metadata.business_assets as ba, dg_metadata.business_assets_type as bat where ba.business_assets_type_id = bat.id and bat.name='TERM' and ba.id IN({{ids}})")(result)
+    val result: String = additionalBusiness.getBusinessTermsPartialIndexationSubquery2("dg_metadata", "business_assets", "business_assets_type", "bpm_status")
+    assertResult("select ba.id as id,ba.name as name,'' as alias,ba.description as description,'' as metadata_path,'GLOSSARY' as type,'BUSINESS_TERM' as subtype,ba.tenant,null as properties,true as active,ba.modified_at as discovered_at,ba.modified_at as modified_at from dg_metadata.business_assets as ba, dg_metadata.business_assets_type as bat, dg_metadata.bpm_status as bas where ba.business_assets_type_id = bat.id and bat.name='TERM' and ba.business_assets_status_id = bas.id and bas.active = true and ba.id IN({{ids}})")(result)
 
   }